// src/domains/dashboard/Payroll/types/types.ts
export type filterStates = {
    search: string;
    start: number;
    length: number;
    year: number;
    month: number;
};